Tweaked fly through speeds.
Now, the minumum fly through speed is 3 seconds, and the max in 5
seconds. Based on the closest travel position and the furthest, we
calculate how long, between 3 & 5 seconds, the travel time should be,
then set the speed accordingly.

Additionally, if the player is closer than the supposed closest
position, we calculate a value between 0 & 3 seconds. This can happen
if the player doesn’t explore very far before the first trial.